Feat: Organized the store page into modular components
